pyproject.toml
[tool.poetry]
name = "vpp-agent"
version = "0.1.0"
description = ""
authors = []
license = "MIT"
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "3.11.15"
langchain-openai = "^1.1.12"
langchain = "^1.2.13"
pydantic-settings = "^2.13.1"
python-dotenv = "^1.2.2"
rich = "^14.3.3"

[tool.poetry.group.test.dependencies]
pytest-asyncio = "0.24.0"
pytest = "^8.3.4"
setuptools = "^82.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
